# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * City Wallet API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Weather {
  temp: number;
  condition: string;
  feelsLike: number;
}

export interface Location {
  city: string;
  district: string;
}

export interface TimeContext {
  hour: number;
  period: string;
}

export type MerchantDemandStatus =
  (typeof MerchantDemandStatus)[keyof typeof MerchantDemandStatus];

export const MerchantDemandStatus = {
  low: "low",
  normal: "normal",
} as const;

export interface MerchantDemand {
  merchantId: string;
  volume: number;
  status: MerchantDemandStatus;
}

export interface ContextState {
  weather: Weather;
  location: Location;
  time: TimeContext;
  merchantDemand: MerchantDemand;
  userBehavior: string;
}

export interface Offer {
  id: string;
  timestamp: string;
  merchantId: string;
  merchantName: string;
  merchantEmoji: string;
  distanceMeters: number;
  headline: string;
  subtext: string;
  discountPercent: number;
  expiryMinutes: number;
  tone: string;
  ctaText: string;
  contextSnapshot: ContextState;
  accepted: boolean;
}

export interface RedeemRequest {
  offerId: string;
  userId: string;
}

export interface RedeemResponse {
  success: boolean;
  token: string;
  qrDataUrl: string;
  cashbackAmount: number;
}

export interface ValidateRedemptionRequest {
  token: string;
}

export interface RedemptionValidation {
  valid: boolean;
  message: string;
}

export type MerchantRulesPreferredTone =
  (typeof MerchantRulesPreferredTone)[keyof typeof MerchantRulesPreferredTone];

export const MerchantRulesPreferredTone = {
  warm: "warm",
  urgent: "urgent",
  playful: "playful",
} as const;

export interface MerchantRules {
  maxDiscountPercent: number;
  triggerStartHour: number;
  triggerEndHour: number;
  volumeThreshold: number;
  preferredTone: MerchantRulesPreferredTone;
  autoGenerate: boolean;
}

export interface Merchant {
  id: string;
  name: string;
  address: string;
  emoji: string;
  distanceMeters: number;
  rules: MerchantRules;
}

export interface OfferPerformance {
  offerId: string;
  headline: string;
  discountPercent: number;
  timestamp: string;
  accepted: boolean;
  acceptRate: number;
  revenue: number;
}

export interface MerchantDashboard {
  merchant: Merchant;
  offersToday: number;
  acceptRate: number;
  totalRevenueLift: number;
  avgDiscount: number;
  currentOffer?: Offer | null;
  viewsProgress: number;
  redemptionsProgress: number;
  offersList: OfferPerformance[];
  context: ContextState;
}

export interface UpdateRulesResponse {
  success: boolean;
  rules: MerchantRules;
}