# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: City Wallet API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /redeem:
    post:
      operationId: redeemOffer
      tags: [redeem]
      summary: Redeem an offer
      description: Generates a QR code for the offer and records the redemption
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedeemRequest"
      responses:
        "200":
          description: Redemption response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedeemResponse"
  /redeem/validate:
    post:
      operationId: validateRedemption
      tags: [redeem]
      summary: Validate a redemption token
      description: Marks a redemption as used after counter scan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateRedemptionRequest"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedemptionValidation"
...
components:
  schemas:
...
    RedeemRequest:
      type: object
      properties:
        offerId:
          type: string
        userId:
          type: string
      required:
        - offerId
        - userId
    RedeemResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
        qrDataUrl:
          type: string
        cashbackAmount:
          type: number
      required:
        - success
        - token
        - qrDataUrl
        - cashbackAmount
    ValidateRedemptionRequest:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    RedemptionValidation:
      type: object
      properties:
        valid:
          type: boolean
        message:
          type: string
      required:
        - valid
        - message
...