# lib/api-zod/src/generated/types/redeemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * City Wallet API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RedeemRequest {
  offerId: string;
  userId: string;
}