# lib/api-zod/src/generated/types/redeemResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * City Wallet API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RedeemResponse {
  success: boolean;
  token: string;
  qrDataUrl: string;
  cashbackAmount: number;
}